refactor: move functions to baseElement
