+++ demo/index.html
@@ -25,7 +25,19 @@
 
       <demo-snippet>
         <template>
+          <auro-input></auro-input>
+        </template>
+      </demo-snippet>
+
+      <demo-snippet>
+        <template>
           <auro-input type="password" label="Password" helptext="Please enter your password" required id="passwordTest01"></auro-input>
+        </template>
+      </demo-snippet>
+
+      <demo-snippet>
+        <template>
+          <auro-input type="email" label="Email Address" helptext="Please enter name@domain.com" required id="emailTest01"></auro-input>
         </template>
       </demo-snippet>
