+++ docs/api.md
@@ -1,7 +1,5 @@
 # auro-input
 
-auro-input provides users a way to enter data into a text field.
-
 ## Properties
 
 | Property                  | Attribute                 | Type      | Description                                      |
@@ -16,9 +14,3 @@ auro-input provides users a way to enter data into a text field.
 | `required`                | `required`                | `Boolean` | Populates the `required` attribute on the input. Used for client-side validation. |
 | `type`                    | `type`                    | `String`  | Populates the `type` attribute on the input. Allowed values are `password`, `email` or `text`. If given value is not allowed or set, defaults to `text`. |
 | `value`                   | `value`                   | `String`  | Populates the `value` attribute on the input. Can also be read to retrieve the current value of the input. |
-
-## Methods
-
-| Method             | Type                         |
-|--------------------|------------------------------|
-| `showPasswordIcon` | `(): TemplateResult \| null` |
